Implements variables updates via a transform
